Fix datetime casting macro for ClickHouse
Previously, when a `NULL` value was encountered in the `timestamp_field`, the `toString` function would convert this NULL value to a `Nullable(String)`. This would result in the split function not working properly, since it expects non-nullable input values. The datetime parsing can be simplified by using `parseDateTimeBestEffortOrNull` instead of doing manual parsing of the date, bypassing the issue of `Nullable` types entirely.